Fixed a bunch of potential issues on 64 bit OS's
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@358 01821946-851b-0410-b7f4-f1477c70e57a

+++ arch/linux/ipx_udp.c
@@ -1177,7 +1177,7 @@ static int UDPReceivePacket (
 if (0 > (dataLen = recvfrom (s->fd, outBuf, outBufSize, 0, (struct sockaddr *) &fromAddr, &fromAddrSize))) {
 		return -1;
 	}
-bTracker = IsTracker (* ((unsigned long *) &fromAddr.sin_addr), * ((ushort *) &fromAddr.sin_port));
+bTracker = IsTracker (* ((unsigned int *) &fromAddr.sin_addr), * ((ushort *) &fromAddr.sin_port));
 #if UDPDEBUG
 	//msg ("ReceivePacket, dataLen=%d", dataLen);
 	//printf (MSGHDR "recvfrom ((%d-8=%d),",dataLen,dataLen-8);
