+++ arch/sdl/rbaudio.c
@@ -225,7 +225,7 @@ static int cddb_sum(int n)
 }
 
 
-unsigned long RBAGetDiscID()
+unsigned int RBAGetDiscID()
 {
 	int i, t = 0, n = 0;
